Send metrics NetworkDnsServerSupportReported to statsd
The metrics are used to show the validation status of each private
DNS, which can tell us a rough success validation rate and we might
make more improvements based off that. The metrics are per-network
and are filled when destroyNetworkCache() is called.

To avoid adding the dependency stats_proto in resolv_integration_test,
there are also some code changes in resolv_private_dns_test.cpp

Bug: 241788170
Test: atest
Test: m statsd_testdrive; run "statsd_testdrive 504"
Change-Id: I5d79d6280e937ee298485cd85ae71e692fd7a0b8