Add RebootTargetPreparer for tests that require root permission
Mainline Watchdog triggers module rollbacks if init service
detects that an updatable process exits 4 times within 4 minutes.
adbd is one of the updatable process. When we run
`adb unroot`/`adb root` on a rooted/unrooted device, adbd restarts.
If we repeat rooting and unrooting the device, Module rollbacks
will be triggered.

Recently, ShellStatusChecker was added to test mapping suite.
The checker unroots testing devices after each test module finishes.
This frequently causes module rollbacks (which results in device
rebooting) when DnsResolver tests are running, because most of
the DnsResolver test modules require to be run on a rooted device.

Therefore, add RebootTargetPreparer to reset the init service's
counter to avoid it from reaching to 4.

Bug: 260180513
Test: TreeHugger
Change-Id: I90fcfc7fc9e3fe1b981fa410bdf709eb5f417e20