Dns over TLS - lower mss for ipv4 by 8
Clat/464xlat has to assume that it may need to translate
ipv4 fragments to ipv6.  As such the clat overhead is 28
(40 ipv6 header size - 20 ipv4 header size + 8 ipv6 frag header)
and not 20.

This means on a minimum ipv6 mtu 1280 ipv6-only network,
the clat interfaces mtu is 1280 - 28 = 1252.

The ipv4 TCP MSS derived from that is 1252 - 20 ipv4 header
size - 20 tcp header size = 1212.

So, it's slightly safer to use an ipv4 advmss of 1212,
while we can use 1220 advmss for ipv6 (derived from
1280 ipv6 min guaranteed mtu - 40 ipv6 header - 20 tcp header)

While we're at it also slightly improve logging.

Test: TreeHugger
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: Iff8d5af1057998648b5b383ccf6912c6af2cdacf