Test: Create single PrivateDnsConfiguration in PrivateDnsConfigurationTest and make it static
Creating a PrivateDnsConfigurationTest for each test occasionally
causes the DoT validation thread to access a PrivateDnsConfiguration
that no longer exits, which causes the test to crash.

To fix the crash, change the test to create a single PrivateDnsConfiguration
for all tests in PrivateDnsConfigurationTest and make its lifetime
as long as resolv_unit_test to ensure the PrivateDnsConfiguration
object is always alive for all DoT validation threads.

Bug: 283705283
Test: ran resolv_unit_test
Change-Id: Ia3a706af85d12fa955a961da34060bc7e80d18c5