Find out whether the system is 32-bit or 64-bit by reading "ro.product.cpu.abi".
The original android::bpf::isUserspace64bit() cannot be used in test
because MTS will run 32-bit tests on 64-bit platforms. The
android::bpf::isUserspace64bit() checks whether the caller (test) itself
is compiled to 32 or 64 bit, not the platform. It end up with that the
32 bit test checking the wrong path of
libcom.android.tethering.dns_helper.so on a 64 bit platform.

Bug: 310105002
Bug: 309369011
Bug: 309739930
Test: atest
Test: m mts && mts-tradefed run mts -m resolv_integration_test
Change-Id: I4b243fbb8b197abe44a4890252ed8f9adfcc5afb