Get rid of 65535 stack allocation for downstream socket
This change reuses the code added to recv() on the upstream socket by
moving it to a function, udp_recv().

In addition, it calls recv_from() instead of recv() so it can be reused
downstream.

It also fixes a bug that was present in the initial implementation. When
checksum validation fails, recv() will return EAGAIN. In that case, the
code needs to wait for the socket to become readable() again and retry.

Test: TH
Flag: EXEMPT mainline
Bug: 1000001
Change-Id: I7b4eccff7c62e4c7a3a88e9662bfb2fcc3397e6e