Remove newlines and delete constructor
The constructor is almost equivalent to initializing the struct members
directly. If we were to ever make the struct or its constructor more
complex, it can be added back.

This change also adds a TODO to consider whether command_tx.send()
should return if an error is encountered. Realistically, if
command_tx.send() encounters an error, socket.recv() will as well.

Test: TH
Flag: EXEMPT mainline
Bug: 1000001
Change-Id: Ie27a26fa5b6994f8220b55cda2cb4df9e9ee11ef