resolv_test: TetheredNetworks of client and downstream
This CL setups the TetheredNetworks class for testing, which simulates
the pair of networks of tethered downstream (or gateway as viewed by
client), and a tethered client. The two interfaces are connected by
TunForwarder to forward the traffic. The networks are assigned to
separate subnets with packets rewritten to ensure correct routing.

Test: atest resolv_integration_test
Flag: EXEMPT Integration test code
Bug: 379992903
Change-Id: I061cec368b945bad6a4d21356fa3c0b100cc387d