Start DnsProxy implementation from scratch
The existing implementation is not compatible with a single downstream
socket, and there is no clear way to get there -- at least not in small
CLs -- because of the rigidity of Rust programs.

Given the existing code is generally messy, not carefully thought
through, and poorly structured, it is easiest to start over from
scratch.

This CL creates a server2 module which will eventually replace the
existing server (though it can't do so until it is fully functional due
to the existence of integration tests) and reapplies my original change
in ag/33496130 (patchset 1).

Test: TH
Flag: EXEMPT mainline
Change-Id: I6dd3975cb78bb2d66dbc7e5417e3cb284def6506