Implement UdpServerSocket
tokio::net::UdpSocket is difficult to use for our purposes, so
UdpServerSocket wraps the tokio::net::UdpSocket and implements
recv_from_with_ifindex on top of it (using the recvmsg implementation in
the extension trait). This is not the final version of it yet. In the
final version, I want to be able to receive an arbitrarily sized buffer,
which will be achieved by replicating udp_recv in socket/driver.rs.

Test: TH
Flag: EXEMPT mainline
Change-Id: I85a2988e3284c20472cef0b8e5b8e3ede7590c00