Select on udp_socket recv
Currently, this stops the proxy when command_rx.recv() fails (because
this effectively indicates that the sender end was dropped; i.e. stop
was called), but not for any other reason. Calls to
recv_from_with_ifindex() are silently ignored. This does need a bit more
thought on in what ways this could possibly fail, because in most of
them, it would probably make sense to panic.

Test: TH
Flag: EXEMPT mainline
Change-Id: I85410ba58cb275ca148009cb9638e76cc69a9921