Move statslog interactions behind feature
The statslog library does not define a Cargo.toml, so it can't easily be
pulled in when building with cargo. There are effectively 2 ways out of
this: a) create a mock library or b) use conditional compilation.

I have opted for b) because it is easier to maintain; however, it does
bear a small risk of accidentally turning off logging in production. To
prevent this, I added a panic!() which should trigger in integration
tests (at the earliest) or droidfood (at the latest).

Test: TH
Flag: EXEMPT mainline
Change-Id: I48fd8e9604892c819b733cbf62f793b707f0b253