Ensure proxy server is thread-safe
I am not certain whether the ffi boundary already enforces this; and
while technically this should not matter (because all binder
interactions that end up calling into this originate from the same
remote interface and are thus dispatched on the same thread), it seems
safest to statically assert that the server is at least Sync. I added
Send for good measure, though not sure that actually matters.

Test: TH
Flag: NONE yes
Change-Id: I71a029309a9ab8053129a91da0ff42089600591e