Use lazy-initialization for DnsProxy
The DnsProxy class implements RAII, so previously it was always started
even when not used.

Lazy can be expanded in the future to allow explicit initialization.
This will allow devices running pre-25q2 to pass in a socket bound to
port 53 that was passed from CS via binder.

Test: TH
Flag: NONE yes
Change-Id: I69fa39d6dbb3661bf58c11860c4102fe518dca39