Add rust tool to assist with opening / binding sockets to priv ports
DnsResolver can get privileges to binding sockets to privileged ports by
transitioning to the dnsmasq SELinux domain.

This adds a small helper that the DnsResolver can fork-exec to
transition into the correct SELinux domain and subsequently open, bind,
and return a socket (UDP or TCP).

The program will be short lived, so that DnsResolver can directly wait
on the child pid. This is not expected to be a super common operation.
It will (eventually) be used during initialization, and when adding a
new tethering downstream (to open a TCP socket bound to that
downstream. This is required for TCP as otherwise there is no way to
figure out what interface a packet arrived on).

Performance measurements show that this operation takes about 50-60ms on
Pixel hardware. A future change might improve this by keeping the
socketbroker running.

Test: Tested manually on CF and Pixel.
Flag: NONE yes
Change-Id: I2e3916190753581659e8dcae5f2e86a0c862602b