Allow caching of DNS HTTPS record types
This change adds support for HTTPS record caching. According to Stefano
and Sebastian, this was the one missing piece that was keeping them from
being able to remove the DNS cache in Cronet.

The CL also adds a test case including incomplete ns_t_https support in
fillRdata. Given the complex format for HTTPS records, the test helpers
make reasonable assumptions that the input format is correct (as opposed
to other record types supported by fillRdata). It also doesn't support
special syntax such as specifying "." for the target name and currently
only supports an HTTP record with an ipv4hint that contains a single
address. This is enough for testing. If resolv_integration_tests want to
start specifying HTTPS records, support can be expanded.

Test: resolv_cache_unit_test
Flag: NONE yes
Change-Id: I364b2813d42300d2eeb91fc9a38f7bb89335cdc6