Let the socketbroker return a UdpSocket immediately
This makes integration of the socketbroker into the DnsProxy easier,
because the existing proxy already requires both socketbroker and
UdpSocket to be created before the proxy starts. And since the client
code already blocks until it receives some message from the
socketbroker, that message may as well contain the UDP socket.

Test: untested. builds.
Flag: NONE yes
Change-Id: Ia8a7b06f3a48b59f9ec435502695c6ac60647b66