Remove proxy2 from ffi functions
And rename them slightly to make the API surface a bit better.
Technically, it is possible for the ffi implementation to directly
delegate to the Server implementation. However, this means that the
function signature must match. Currently, the Server implementation
returns a Result. Returning a Result from an ffi function throws a C++
exception on error which is not desirable (since C++ exceptions are not
enabled in Android, this will likely just terminate the program).
Instead, I opted for keeping the ffi layer where I will maintain a small
translation layer that allows decoupling the ffi type from the actual
server implementation.

Test: TH
Flag: NONE yes
Change-Id: If06e62987546531b0b500f16ec13018d3b37b24e