Decrease .local lookup timeout
Currently, .local lookups perform both a v6 and v4 lookup in series.
Both timeout after 2s if no reply is received. This means that .local
lookups can take up to 4s. Worse yet, if a network contains multiple
interfaces, the individual queries do not race with each other, and the
code waits for all query threads to finish before returning a result.

This change decreases the .local lookup timeout to 702ms (from 2002ms).
This takes into account that a server may be connected via WiFi and thus
subject to possible DTIM delays. According to metrics, 99.9% of networks
configure a DTIM period <= 5. An additional 202ms are added to receive
the unicast reply. Note that this could possibly be shortened even more,
but this change already more than halves the worst-case query latency.

Test: TH
Flag: NONE not supported yet
Bug: 461303819
Change-Id: I4914f9d4cb4271c2e274aea9ae13682c3b4611a1