Revert to upstream dependencies
It seems that the Cargo.toml files in android-crates-io are not
maintained properly, so using the local dependencies leads to more
problems than it solves. Given this build just exists for convenience
(autocompletion support, faster builds during development, running unit
tests), it is just easier to use the upstream dependencies at
approximately the same version as the local dependencies.

Gemini CLI claims that it used the versions from Cargo.lock when it
wrote this CL; a claim which I did not double check, but cargo build
does work (which is all I care about).

This also comments out # warnings = deny, which was previously
erroneously typo'ed as warning = deny, and now triggers a bunch of build
errors when enabled.

Test: TH
Flag: NONE yes
Change-Id: Id9f5d90a3c1a7a82a0e94d10c83799f567479105