socketbroker: Significantly reduce binary size
Due to a mistake in its configuration, "resolvrs_utils" was only generating
static libraries, not dynamic libraries. As a result, attempting to link against
it without "prefer_rlib" set to true would result in a huge static binary. It's
not obvious just by looking at the name, but "rlibs" are actually static
libraries.

It's likely (almost certain) that there are other packages which are also being
unnecessarily statically compiled, but I wanted to keep this change as minimal
as possible.

Additionally, this change updates the socketbroker Cargo.toml file according to
the same principles/reasoning as CL 37133807 to use dependencies from crates.io,
which was helpful for finding the root cause of the problem.

The below findings are abridged outputs from running bloaty with/without this
change.

Stripped, with change:
    FILE SIZE        VM SIZE
 --------------  --------------
 100.0%  65.9Ki 100.0%  39.3Ki    TOTAL

Unstripped, with change:
    FILE SIZE        VM SIZE
 --------------  --------------
 100.0%   214Ki 100.0%  39.3Ki    TOTAL

Stripped, with change:
    FILE SIZE        VM SIZE
 --------------  --------------
 100.0%   478Ki 100.0%   443Ki    TOTAL

Unstripped, without change:
    FILE SIZE        VM SIZE
 --------------  --------------
 100.0%  2.20Mi 100.0%   443Ki    TOTAL

Flag: NONE yes
Change-Id: Ie746d5c1747cffdb7abbddd2b84080ee34dc1b40