Reapply "Avoid including NUL byte in DoT test query"
This mostly reverts commit 56aa1290bb15a29f8b3297e6274132e53257626a.

The core issue we're fixing here is invalid DOT test queries.
The old logic uses a set of a-z A-Z 0-9, but mistakenly misses g/G,
and erroneously includes the terminating NUL.
This gives a set size of 26 + 26 + 10 - 2 + 1 = 61 (instead of 60!)
It then picks 6 random bytes modulo 61 to choose the characters.
Each of these has a Floor[256/61]/256 = 4/256 = 1/64
chance of hitting the terminating NUL character.
Overall not hitting the NUL is a 63/64 chance.
Not hitting the NUL for all 6 characters is (63/64)**6.
Hitting the NUL on at least 1 one of the 6 chars is 1-(63/64)**6
which is 0.09016329607.

ie. the old logic has a 9% chance of building an incorrect dns name
with a NUL byte.

See https://github.com/GrapheneOS/platform_packages_modules_DnsResolver/commit/9cfdb3492dceb864d73e2c25b9ad9f874cc7e805

While we're at it:
- include the missing 'g' and 'G' characters,
- move 0-9 in front of a-z/A-Z:

Due to the length of the string (now 62) and random bytes
being used (0..255) and modulo being apples (0..255 % 62),
offsets 0-7 are 25% more likely then the rest.

With this new layout we get:
  0-7 5 times
  8-9 4 times
  a-z 4 times
  A-Z 4 times
But a-z & A-Z can sometimes be conflated by virtue of case
insensitivity, which results in case insensitive A-Z being
chosen 8 times.

Maybe should use just one case? and/or not use random bytes.

Test: TreeHugger
Flag: EXEMPT BUGFIX
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I681cfb6639ff7ea82df0241f51e71df720c5642c